+++ test/integration/data/db-compaction.js
@@ -16,6 +16,8 @@ describe(
       const ls = exec('node ' + procPath);
 
       ls.stderr.on('data', data => {
+        //skip deprecation warning
+        if (data.indexOf('Transform.prototype._transformState is deprecated') > -1) return;
         //eslint-disable-next-line
         console.log('stderr:::', data);
         done(data);